remove utils and move to validation package
