+++ src/main/java/validation/ConsoleInputHandler.java
@@ -1,4 +1,4 @@
-package utils;
+package validation;
 
 import java.util.Scanner;
