+++ src/main/java/validation/InputValidationUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package validation;
 
 import CustomException.ValidationException;
